backend_service/requirements.txt
boto3==1.37.33
botocore==1.37.33
PyJWT==2.10.1
python-dotenv==1.1.0
bcrypt==4.3.0
watchtower==3.4.0
